[script] add: Context::preprocessed_
This contains the entire flattened include stack as it might look if you had implicit ifndefs in nwscript header files.  Use it for all symbol look ups.
This ensures that, like a C preprocessed file a header can see things included before it but not explicit in its own dependancy chain.